Fix wrong data type

+++ Team17/Code17/src/spa/src/components/sp/extractor/ProcedureStack.cpp
@@ -21,7 +21,7 @@ void ProcedureStack::close(int statementNumber) {
 void ProcedureStack::mergeStack() {}
 
 void ProcedureStack::extractFollows(std::vector<SimpleToken> follows) {
-    for (int i = 0; i < follows.size() - 1; i++) {
+    for (size_t i = 0; i < follows.size() - 1; i++) {
         SimpleToken first = follows.at(i);
         SimpleToken second = follows.at(i + 1);
         Entity* firstEntity = generateEntity(first);
